+++ NEWS.md
@@ -1,5 +1,9 @@
-# scholar 0.2.2.991
+# scholar 0.2.3
 
++ `format_publications` to format publication list (2022-06-21, Tue)
+    - <https://github.com/jkeirstead/scholar/issues/110>
++ update journal ranking data
++ remove `get_impactfactor`
 + fixed when some years contain 0 cites (@jefferis, #101)
 + update documents (@jefferis, #100)